clean up files

+++ build/windows/Makefile.ps1
@@ -18,14 +18,10 @@ if ($false -eq (Test-Path -Path $builddir)) {
     Write-Host("Invalid build dir : " + $builddir + " ") -ForegroundColor Red
     exit
 }
-$srcdir = Split-Path -Path $builddir
-if ($false -eq (Test-Path -Path $srcdir)) {
-    Write-Host("Invalid docker provider root source dir : " + $srcdir + " ") -ForegroundColor Red
-    exit
-}
-$certsrcdir = Join-Path -Path $srcdir -ChildPath "kubernetes\windows\certificategenerator"
+
+$certsrcdir = Join-Path -Path $builddir -ChildPath "windows\installer\certificategenerator"
 if ($false -eq (Test-Path -Path $certsrcdir)) {
-    Write-Host("Invalid certificate generator source dir : " + $srcdir + " ") -ForegroundColor Red
+    Write-Host("Invalid certificate generator source dir : " + $certsrcdir + " ") -ForegroundColor Red
     exit
 }
 Write-Host("set the cerificate generator source code director : " + $certsrcdir + " ...")
@@ -47,7 +43,14 @@ if ($false -eq (Test-Path -Path $certreleasebinpath)) {
     Write-Host("certificate release bin path doesnt exist : " + $certreleasebinpath + " ") -ForegroundColor Red
     exit
 }
-$publishdir = Join-Path -Path $srcdir -ChildPath "kubernetes\windows\omsagentwindows"
+
+$rootdir = Split-Path -Path $builddir
+if ($false -eq (Test-Path -Path $rootdir)) {
+    Write-Host("Invalid docker provider root source dir : " + $rootdir + " ") -ForegroundColor Red
+    exit
+}
+
+$publishdir = Join-Path -Path $rootdir -ChildPath "kubernetes\windows\omsagentwindows"
 if ($true -eq (Test-Path -Path $publishdir)) {
     Write-Host("publish dir exist hence deleting: " + $publishdir + " ")
     Remove-Item -Path $publishdir  -Recurse -Force
@@ -62,7 +65,7 @@ Compress-Archive -Path  $certreleasebinpath -DestinationPath $certreleasepublish
 Write-Host("Successfully copied compressed certificate generator release binaries") -ForegroundColor Green
 
 # build  the shared object (.so) for out oms go plugin code
-$outomsgoplugindir = Join-Path -Path $srcdir -ChildPath "source\code\go\src\plugins"
+$outomsgoplugindir = Join-Path -Path $rootdir -ChildPath "source\code\go\src\plugins"
 Write-Host("Building Out_OMS go plugin code...")
 if ($false -eq (Test-Path -Path $outomsgoplugindir)) {
     Write-Host("Invalid Out oms go plugin code dir : " + $outomsgoplugindir + " ") -ForegroundColor Red
@@ -78,9 +81,12 @@ Write-Host("Successfully build Out_OMS go plugin code") -ForegroundColor Green
 Write-Host("copying out_oms.so file to : $publishdir")
 Copy-Item -Path (Join-path -Path $outomsgoplugindir -ChildPath "out_oms.so")  -Destination $publishdir -Force
 Write-Host("successfully copied out_oms.so file to : $publishdir") -ForegroundColor Green
-Set-Location $currentdir
+
 
 $installerdir = Join-Path -Path $builddir -ChildPath "windows\installer"
 Write-Host("copying installer files conf and scripts from :" + $installerdir + "  to  :" + $publishdir + " ...")
-Copy-Item  -Path $installerdir  -Destination $publishdir -Recurse -Force
+$exclude = @('*.cs','*.csproj')
+Copy-Item  -Path $installerdir  -Destination $publishdir -Recurse -Force -Exclude $exclude
 Write-Host("successfully copied installer files conf and scripts from :" + $installerdir + "  to  :" + $publishdir + " ") -ForegroundColor Green
+
+Set-Location $currentdir
